Fixing export with format filters and fixing checkboxes for IE
