+++ Catalogue.Data/Write/RecordValidator.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Text.RegularExpressions;
-using Catalogue.Data.Model;
+﻿using Catalogue.Data.Model;
 using Catalogue.Gemini.Helpers;
 using Catalogue.Gemini.Model;
 using Catalogue.Gemini.ResourceType;
@@ -16,8 +10,11 @@ using Catalogue.Utilities.Clone;
 using Catalogue.Utilities.Collections;
 using Catalogue.Utilities.Text;
 using FluentAssertions;
-using Moq;
 using NUnit.Framework;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.RegularExpressions;
 
 namespace Catalogue.Data.Write
 {
@@ -277,7 +274,7 @@ namespace Catalogue.Data.Write
                 result.Errors.Add("MESH GUI not valid", r => r.Gemini.Keywords);
         }
 
-        internal static bool IsValidDate(string date)
+        public static bool IsValidDate(string date)
         {
             //todo: Disabled for mest data testing.
             var yearOnly = new Regex(@"^\d\d\d\d$");
@@ -464,336 +461,5 @@ namespace Catalogue.Data.Write
                 }
             }
         }
-    }
-
-
-    internal class when_validating_at_basic_level
-    {
-        Record SimpleRecord()
-        {
-            return new Record
-            {
-                Path = @"X:\some\path",
-                Gemini = Library.Blank().With(m =>
-                    {
-                        m.Title = "Some title";
-                        m.Keywords = new StringPairList
-                            {
-                                { "http://vocab.jncc.gov.uk/jncc-domain", "Example Domain" },
-                                { "http://vocab.jncc.gov.uk/jncc-category", "Example Category" },
-                            }
-                            .ToKeywordList();
-                    }),
-            };
-        }
-
-        [Test]
-        public void should_produce_no_warnings_by_default()
-        {
-            // the basic level of validation shouldn't produce warnings - that would be too annoying
-
-            var result = new RecordValidator().Validate(SimpleRecord() /* no Level argument */);
-            result.Warnings.Should().BeEmpty();
-        }
-
-        [Test]
-        public void title_must_not_be_blank([Values("", " ", null)] string blank)
-        {
-            var result =
-                new RecordValidator().Validate(SimpleRecord().With(r => r.Gemini.Title = blank));
-
-            result.Errors.Single().Message.Should().StartWith("Title must not be blank");
-            result.Errors.Single().Fields.Single().Should().Be("gemini.title");
-        }
-
-        [Test]
-        public void title_must_not_be_londer_then_200([Values("", " ", null)] string blank)
-        {
-            var result =
-                new RecordValidator().Validate(SimpleRecord().With(r => r.Gemini.Title = new String('x', 201)));
-
-            result.Errors.Single().Message.Should().StartWith("Title is too long. 200 characters or less, please");
-            result.Errors.Single().Fields.Single().Should().Be("gemini.title");
-        }
-
-        [Test]
-        public void publication_title_must_not_be_londer_then_250([Values("", " ", null)] string blank)
-        {
-            var result =
-                new RecordValidator().Validate(SimpleRecord().With(r => { r.Gemini.Title = new String('x', 251);
-                                                                            r.Gemini.ResourceType = "publication";
-                }));
-
-            result.Errors.Single().Message.Should().StartWith("Title is too long. 250 characters or less, please");
-            result.Errors.Single().Fields.Single().Should().Be("gemini.title");
-        }
-
-        [Test]
-        public void publication_title_can_be_londer_then_150([Values("", " ", null)] string blank)
-        {
-            var result =
-                new RecordValidator().Validate(SimpleRecord().With(r =>
-                {
-                    r.Gemini.Title = new String('x', 155);
-                    r.Gemini.ResourceType = "publication";
-                }));
-
-            result.Errors.Count().Should().Be(0);
-        }
-
-        [Test]
-        public void path_must_not_be_blank([Values("", " ", null)] string blank)
-        {
-            var result = new RecordValidator().Validate(SimpleRecord().With(r => r.Path = blank));
-
-            result.Errors.Single().Fields.Single().Should().Be("path");
-        }
-
-        [Test]
-        public void path_must_be_valid()
-        {
-            var result = new RecordValidator().Validate(SimpleRecord().With(r => r.Path = "not a valid path"));
-            result.Errors.Single().Fields.Single().Should().Be("path");
-        }
-
-        [Test]
-        public void path_must_be_an_acceptable_kind([Values(@"X:\some\path", "PG:\"host=spatial-store dbname=spatial layer=SSSI_England_Units\"")] string path)
-        {
-            var result = new RecordValidator().Validate(SimpleRecord().With(r => r.Path = path));
-            result.Errors.Should().BeEmpty();
-        }
-
-        [Test]
-        public void jncc_keywords_must_be_provided()
-        {
-            // should not validate on empty list
-            var r1 =
-               new RecordValidator().Validate(SimpleRecord().With(r => r.Gemini.Keywords = new List<MetadataKeyword>()));
-
-            r1.Errors.First().Message.Should().StartWith("Must specify a JNCC Domain");
-        }
-
-        [Test]
-        public void keywords_may_not_be_blank()
-        {
-            var record = SimpleRecord();
-            record.Gemini.Keywords.Add(new MetadataKeyword() { Value = String.Empty, Vocab = String.Empty });
-
-            var result =
-                new RecordValidator().Validate(record);
-
-            result.Errors.Single().Message.Should().StartWith("Keywords cannot be blank");
-            result.Errors.Single().Fields.Single().Should().Be("gemini.keywords");
-        }
-
-        [Test]
-        public void dataset_reference_date_must_be_valid_date([Values("", " ", null)] string invalid)
-        {
-            var validDates = new[]
-                {
-                    "",
-                    "",
-                    "",
-                    "",
-                };
-//            var result = new RecordValidator().Validate(SimpleRecord().With(r => r.Gemini.DatasetReferenceDate));
-//
-//            result.Errors.Single().Fields.Single().Should().Be("path");
-        }
-
-        [Test]
-        public void topic_category_must_be_valid()
-        {
-            var record = SimpleRecord().With(r => r.Gemini.TopicCategory = "anInvalidTopicCategory");
-            var result = new RecordValidator().Validate(record);
-
-            result.Errors.Single().Message.Should().Contain("Topic Category 'anInvalidTopicCategory' is not valid");
-            result.Errors.Single().Fields.Single().Should().Be("gemini.topicCategory");
-        }
-
-        [Test]
-        public void topic_category_may_be_blank([Values("", null)] string blank)
-        {
-            var record = SimpleRecord().With(r => r.Gemini.TopicCategory = blank);
-            var result = new RecordValidator().Validate(record);
-            result.Errors.Should().BeEmpty();
-        }
-
-        [Test]
-        public void sensitive_records_must_have_limitations_on_public_access(
-            [Values(Security.Secret, Security.OfficialSensitive)] Model.Security nonOpen, [Values("", " ", null)] string blank)
-        {
-            var record = SimpleRecord().With(r =>
-            {
-                r.Security = nonOpen;
-                r.Gemini.LimitationsOnPublicAccess = blank;
-            });
-
-            var result = new RecordValidator().Validate(record);
-
-            result.Errors.Single().Fields.Should().Contain("security");
-            result.Errors.Single().Fields.Should().Contain("gemini.limitationsOnPublicAccess");
-        }
-
-        [Test]
-        public void resource_locator_must_be_a_well_formed_http_url(
-            [Values(@"Z:\some\path", "utter rubbish")] string nonHttpUrl)
-        {
-            var record = SimpleRecord().With(r => r.Gemini.ResourceLocator = nonHttpUrl);
-            var result = new RecordValidator().Validate(record);
-            result.Errors.Single().Fields.Should().Contain("gemini.resourceLocator");
-        }
-
-        [Test]
-        public void resource_locator_may_be_set()
-        {
-            var record = SimpleRecord().With(r => r.Gemini.ResourceLocator = "http://example.org/resource/locator");
-            var result = new RecordValidator().Validate(record);
-            result.Errors.Should().BeEmpty();
-        }
-
-        [Test]
-        public void responsible_organisation_role_must_be_an_allowed_role()
-        {
-            var record = SimpleRecord().With(r => r.Gemini.ResponsibleOrganisation = new ResponsibleParty
-            {
-                Email = "a.mann@example.com",
-                Name = "A. Mann",
-                Role = "some role that isn't allowed",
-            });
-            var result = new RecordValidator().Validate(record);
-            result.Errors.Single().Fields.Should().Contain("gemini.responsibleOrganisation.role");
-        }
-
-        [Test]
-        public void metadata_point_of_contact_role_must_be_an_allowed_role()
-        {
-            var record = SimpleRecord().With(r => r.Gemini.MetadataPointOfContact = new ResponsibleParty
-            {
-                Email = "a.mann@example.com",
-                Name = "A. Mann",
-                Role = "some role that isn't allowed",
-            });
-            var result = new RecordValidator().Validate(record);
-            result.Errors.Single().Fields.Should().Contain("gemini.metadataPointOfContact.role");
-        }
-
-        [Test]
-        public void should_validate_valid_dates()
-        {
-            RecordValidator.IsValidDate("2017").Should().BeTrue();
-            RecordValidator.IsValidDate("2017-01").Should().BeTrue();
-            RecordValidator.IsValidDate("2017-01-01").Should().BeTrue();
-            RecordValidator.IsValidDate("2012-02-29").Should().BeTrue(); //leap year
-
-        }
-
-        [Test]
-        public void should_not_validate_invalid_dates()
-        {
-            RecordValidator.IsValidDate("aaa").Should().BeFalse();
-            RecordValidator.IsValidDate("August 2009").Should().BeFalse();
-            RecordValidator.IsValidDate("2013-02-29").Should().BeFalse(); //not a leap year
-        }
-
-        [Test]
-        public void mesh_gui_keywords_must_be_of_valid_form()
-        {
-            var record = SimpleRecord().With(r => r.Gemini.Keywords.Add(new MetadataKeyword {Vocab = "http://vocab.jncc.gov.uk/mesh-gui", Value = "blah"} ));
-            var result = new RecordValidator().Validate(record);
-            result.Errors.Single().Fields.Should().Contain("gemini.keywords");
-        }
-    }
-
-    internal class when_validating_at_gemini_level
-    {
-        private Record GeminiRecord()
-        {
-            return new Record
-            {
-                Path = @"X:\some\path",
-                Gemini = Library.Example().With(m =>
-                {
-                    m.Keywords = new StringPairList
-                            {
-                                { "http://vocab.jncc.gov.uk/jncc-domain", "Example Domain" },
-                                { "http://vocab.jncc.gov.uk/jncc-category", "Example Category" },
-                            }
-                        .ToKeywordList();                    
-                }),
-
-                Validation = Validation.Gemini,
-            };
-        }
-
-        [Test]
-        public void should_validate_valid_gemini_record()
-        {
-            var record = GeminiRecord();
-
-            var result = new RecordValidator().Validate(record);
-
-            result.Errors.Count.Should().Be(0);
-        }
-
-        [Test]
-        public void should_validate_valid_nonspatial_record_with_no_bbox()
-        {
-            // to keep things comceptually simple, let's try to keep one level of "good" validation - "gemini"
-            // but use it for non-spatial metadata too
-
-            var record = GeminiRecord().With(r =>
-                {
-                    r.Gemini.ResourceType = "nonGeographicDataset";
-                    r.Gemini.BoundingBox = null;
-                });
-
-            var result = new RecordValidator().Validate(record);
-
-            result.Errors.Count.Should().Be(0);
-        }
-
-        [Test]
-        public void responsible_organisation_role_must_not_be_blank()
-        {
-            var record = GeminiRecord().With(r => r.Gemini.ResponsibleOrganisation = new ResponsibleParty
-            {
-                Email = "a.mann@example.com",
-                Name = "A. Mann",
-                Role = "",
-            });
-            var result = new RecordValidator().Validate(record);
-            result.Errors.Single().Fields.Should().Contain("gemini.responsibleOrganisation.role");
-        }
-
-        [Test]
-        public void blank_lineage_is_not_allowed([Values("", " ", null)] string blank)
-        {
-            var record = GeminiRecord().With(r => r.Gemini.Lineage = blank);
-            var result = new RecordValidator().Validate(record);
-
-            result.Errors.Any(e => e.Fields.Contains("gemini.lineage")).Should().BeTrue();
-        }
-
-        [Test]
-        public void blank_use_constraints_are_not_allowed([Values("", " ", null)] string blank)
-        {
-            var record = GeminiRecord().With(r => r.Gemini.UseConstraints = blank);
-            var result = new RecordValidator().Validate(record);
-
-            result.Errors.Any(e => e.Fields.Contains("gemini.useConstraints")).Should().BeTrue();
-        }
-
-        [Test]
-        public void topic_category_must_not_be_blank([Values("", " ", null)] string blank)
-        {
-            var record = GeminiRecord().With(r => r.Gemini.TopicCategory = blank);
-            var result = new RecordValidator().Validate(record);
-
-            result.Errors.Any(e => e.Fields.Contains("gemini.topicCategory")).Should().BeTrue();
-        }
-
-//        [Test]
-//        public void should_not_allow_keyword_additions_to_controlled_vocabs()
     }
 }
