+++ Catalogue.Web/app/views/search/search.html
@@ -28,7 +28,7 @@
                 <h4>Advanced search options</h4>
                 <form>
                     <label tc-tip title="Filter by data format (multiple choice)" ng-repeat="f in dataFormatOptions">
-                        <input type="checkbox" ng-checked="query.f.dataFormats.includes(f)" ng-click="addOrRemoveDataFormat(f)"/> {{f}}&nbsp;&nbsp;
+                        <input type="checkbox" ng-checked="dataFormatSelections[f]" ng-click="addOrRemoveDataFormat(f)"/> {{f}}&nbsp;&nbsp;
                     </label>
                 </form>
             </div>
